+++ src/lib.rs
@@ -33,5 +33,8 @@ mod alloc {
 pub use target_lexicon;
 pub use uuid;
 
+mod common;
+pub use common::*;
+
 pub mod read;
 pub use read::*;
